+++ web/src/components/card/ResultCard.tsx
@@ -6,12 +6,37 @@ import Button from "../button/button";
 import Carousel from "../carousel/carousel";
 import LegendAccordion from "../LegendAccordion/LegendAccordion";
 
-const ResultCard: React.FC = () => {
+type ItemType = {
+  title: string;
+  description: string;
+  color: string;
+};
+
+const legendItems: ItemType[] = [
+  {
+    title: "Гарь",
+    description: "Территория, пострадавшая от лесного пожара.",
+    color: "#FF0000",
+  },
+  {
+    title: "Пашни",
+    description: "Земли, используемые для сельскохозяйственных работ.",
+    color: "#FDC611",
+  },
+  {
+    title: "Леса",
+    description: "Натуральный лесной массив, охраняемый от вырубки.",
+    color: "#00CC00",
+  },
+];
+
+interface ResultProps  {
+  previews: string[]
+}
+
+
+const ResultCard: React.FC<ResultProps> = ({previews}: {previews: string[]}) => {
   const cardClass = classNames(styles.card_result);
-  const previewUrls = [
-    "https://i.imgur.com/YzFSzED.jpeg",
-    "https://i.imgur.com/3giN25k.jpeg",
-  ];
 
   return (
     <div className={styles.mainContent}>
@@ -20,19 +45,19 @@ const ResultCard: React.FC = () => {
           <h2 className={styles.headText}>Ваш результат</h2>
 
           <div className={styles.imageWrapper}>
-            <Carousel previewUrls={previewUrls} />
+            <Carousel previewUrls={previews} />
           </div>
 
-          <LegendAccordion></LegendAccordion>
+          <LegendAccordion legendItems={[]} ></LegendAccordion>
 
           <div className={styles.buttonsWrapper}>
             <Button
               size="large"
               type="submit"
               variant="mint"
-              disabled={previewUrls.length === 0}
+              disabled={previews.length === 0}
             >
-              Скачать ({previewUrls.length})
+              Скачать ({previews.length})
             </Button>
           </div>
         </div>
